// src/interfaces/global-variants.d.ts
export declare enum GlobalVariants {
    Zps = "ZPS",
    Cat = "CAT",
    Zsd = "ZSD",
    Zhd = "ZHD"
}
export declare enum GlobalVariantExtended {
    ZpsBg = "ZPS_BG",
    CatBg = "CAT_BG"
}
export declare enum LocaleVariants {
    EN = "en",
    DE = "de",
    FR = "fr",
    GB = "en-GB",
    IT = "it",
    IN = "en-IN",
    US = "en-US",
    ZH = "zh"
}
//# sourceMappingURL=global-variants.d.ts.map